requirements.txt
cs50==9.0.0
emoji==1.7.0
Pillow==9.2.0
pyfiglet==0.8.post1
requests==2.28.0
SQLAlchemy==1.4.36
tabulate==0.8.10
validator_collection==1.5.0